extract DataFormat into separate file
