+++ webknossos/webknossos/dataset/__init__.py
@@ -9,7 +9,7 @@ The dataset stores the data on disk in `.wkw`-files (see [WEBKNOSSOS-wrap (wkw)]
 Each dataset consists of one or more layers (webknossos.dataset.layer.Layer), which themselves can comprise multiple magnifications (webknossos.dataset.mag_view.MagView).
 """
 
-from ._array import DataFormat
+from .data_format import DataFormat
 from .dataset import Dataset, RemoteDataset
 from .layer import Layer, SegmentationLayer
 from .layer_categories import COLOR_CATEGORY, SEGMENTATION_CATEGORY, LayerCategoryType
